+++ packages/firebase-hooks/src/index.js
@@ -2,7 +2,7 @@ export { default as FirebaseProvider } from './firebase-provider';
 export { default as useFirebaseApp } from './useFirebaseApp';
 export { default as useCollectionOnce } from './firestore/useCollectionOnce';
 export { default as useCollection } from './firestore/useCollection';
-export { default as useCurrentUser } from './firestore/useCurrentUser';
 export { default as useDoc } from './firestore/useDoc';
 export { default as useDocOnce } from './firestore/useDocOnce';
 export { default as useAuthState } from './auth/useAuthState';
+export { default as useCurrentUser } from './auth/useCurrentUser';