+++ packages/icons/docs/stories/helpers/iconsArray.js
@@ -63,6 +63,10 @@ import {
     CloseCircleFilled,
     Email,
     QuestionFilled,
+    LocationPin,
+    LocationPinFilled,
+    Download,
+    OpenBookFilled,
 } from '../../../src';
 
 export default [
@@ -386,5 +390,25 @@ export default [
         name: 'questionFilled',
         category: 'standard',
         icon: QuestionFilled,
+    },
+    {
+        name: 'locationPin',
+        category: 'standard',
+        icon: LocationPin,
+    },
+    {
+        name: 'locationPinFilled',
+        category: 'standard',
+        icon: LocationPinFilled,
+    },
+    {
+        name: 'openBookFilled',
+        category: 'standard',
+        icon: OpenBookFilled,
+    },
+    {
+        name: 'download',
+        category: 'standard',
+        icon: Download,
     },
 ];