+++ packages/icons/src/index.js
@@ -65,3 +65,7 @@ export { default as FacebookCircle } from './components/facebookCircle';
 export { default as CloseCircleFilled } from './components/closeCircleFilled';
 export { default as Email } from './components/email';
 export { default as QuestionFilled } from './components/questionFilled';
+export { default as LocationPinFilled } from './components/locationPinFilled';
+export { default as LocationPin } from './components/locationPin';
+export { default as OpenBookFilled } from './components/openBookFilled';
+export { default as Download } from './components/download';