+++ packages/search/docs/stories/GlobalSearch.story.js
@@ -1,8 +1,10 @@
 import React from 'react';
 import styled from 'styled-components';
 import { RainbowFirebaseApp } from '@rainbow-modules/app';
+import { Github, Avatar } from '@rainbow-modules/icons';
 import GlobalSearch from '../../src/components/GlobalSearch';
 import SearchEntity from '../../src/components/SearchEntity';
+import useLocalRecentSearches from '../../src/hooks/useLocalRecentSearches';
 import searchBooksAlgolia from './algolia/searchBooksAlgolia';
 import searchAuthorsAlgolia from './algolia/searchAuthorsAlgolia';
 import searchMoviesTitleMongo from './mongo/searchMoviesTitleMongo';
@@ -26,11 +28,13 @@ export const AlgoliaGlobalSearch = () => {
                     name="Books"
                     onAutocomplete={searchBooksAlgolia}
                     onSearch={searchBooksAlgolia}
+                    icon={<Github />}
                 />
                 <SearchEntity
                     name="Authors"
                     onAutocomplete={searchAuthorsAlgolia}
                     onSearch={searchAuthorsAlgolia}
+                    icon={<Avatar />}
                 />
             </StyledGlobalSearch>
         </RainbowFirebaseApp>
@@ -38,6 +42,7 @@ export const AlgoliaGlobalSearch = () => {
 };
 
 export const MongoAtlasGlobalSearch = () => {
+    const [recents, addRecents] = useLocalRecentSearches('mongo-atlas-recent-searches');
     return (
         <RainbowFirebaseApp>
             <StyledGlobalSearch
@@ -45,6 +50,8 @@ export const MongoAtlasGlobalSearch = () => {
                 placeholder="Search"
                 // eslint-disable-next-line no-alert
                 onSelect={(item) => alert(JSON.stringify(item))}
+                recents={recents}
+                onSearch={({ query }) => addRecents(query)}
             >
                 <SearchEntity name="Movies" onAutocomplete={searchMoviesTitleMongo} />
                 <SearchEntity name="Plot" onAutocomplete={searchMoviesPlotMongo} />